backends-main/requirements.txt
flask
firebase-admin
gunicorn
python-dotenv
Flask-Talisman
Flask-WTF